[Guardian] reorg hashi-guardian crate into ceremony_mode/ + withdraw_mode/
Group the flat handler files by the EnclaveMode axis they're gated on, so the
mode a flow runs in is structural rather than reconstructed from the rpc.rs
gates. Pure moves + path requalification; no logic change.

- ceremony_mode/{setup,rotate}.rs  — CEREMONY_MODE=true flows
- withdraw_mode/{standard,committee_update,provisioner_init,heartbeat}.rs
  — CEREMONY_MODE=false flows; mod.rs holds verify_hashi_cert (shared by
  standard + committee_update)
- attestation.rs (NSM) + info.rs (get_guardian_info) — split from getters.rs
- operator_init.rs — split from init.rs (runs in both modes; provisioner_init
  moved to withdraw_mode/)
- LimiterGuard moved from the withdrawal handler into enclave.rs, next to its
  only constructor EnclaveState::consume_from_limiter (fixes the previous
  enclave -> withdraw back-dependency)

External consumers only use Enclave / S3Logger / rpc / cache / test_utils /
HEARTBEAT_* — none of which moved — so monitor/hashi/e2e compile unchanged.
Hardening cleanups (cache.rs stale comments, list_objects_sample, etc.) are
left for a separate PR.
