+++ crates/hashi-guardian/src/ceremony_mode/rotate.rs
