+++ crates/hashi-guardian/src/ceremony_mode/setup.rs
